Use Rails concerns for Models
Add NodeShared and CommentsShared module for Shared methods

Add notify_callout_users and notify_users method in Comments model
