+++ app/models/concerns/node_shared.rb
@@ -1,0 +1,16 @@
+module NodeShared
+  extend ActiveSupport::Concern
+
+  def likes
+    self.cached_likes
+  end
+
+   def liked_by(uid)
+    self.likers.collect(&:uid).include?(uid)
+  end
+
+  def comments
+    self.drupal_comments
+        .order('timestamp DESC')
+  end
+end
