+++ app/models/drupal_comment.rb
@@ -1,4 +1,6 @@
 class DrupalComment < ActiveRecord::Base
+  include CommentsShared
+
   attr_accessible :pid, :nid, :uid, :aid,
                   :subject, :hostname, :comment,
                   :status, :format, :thread, :timestamp
@@ -41,16 +43,6 @@ class DrupalComment < ActiveRecord::Base
     finder.gsub(Callouts.const_get(:HASHTAG), Callouts.const_get(:HASHLINKMD))
   end
 
-  # filtered version additionally appending http/https
-  #   protocol to protocol-relative URLslike "//publiclab.org/foo"
-  def body_email
-    self.body.gsub(/([\s|"|'|\[|\(])(\/\/)([\w]?\.?publiclab.org)/, '\1https://\3')
-  end
-
-  def author
-    DrupalUsers.find_by_uid self.uid
-  end
-
   def icon
     "<i class='icon-comment'></i>"
   end
@@ -75,10 +67,6 @@ class DrupalComment < ActiveRecord::Base
     end
   end
 
-  def node
-    self.drupal_node
-  end
-
   # users who are involved in this comment thread
   def thread_participants
 
@@ -89,32 +77,38 @@ class DrupalComment < ActiveRecord::Base
     User.find_all_by_username(usernames.map {|m| m[1] }).uniq
   end
 
+  def notify_callout_users
+    # notify mentioned users
+    self.mentioned_users.each do |user|
+      CommentMailer.notify_callout(self,user) if user.username != self.author.username
+    end
+  end
+
+  def notify_users(uids, current_user)
+    DrupalUsers.find(:all, :conditions => ['uid IN (?)',uids]).each do |user|
+      if user.uid != current_user.uid
+        CommentMailer.notify(user.user,self).deliver
+      end
+    end
+  end
+
   # email all users in this thread
   # plus all who've starred it
   def notify(current_user)
     if self.parent.uid != current_user.uid
       CommentMailer.notify_note_author(self.parent.author,self).deliver
     end
-    # notify_callout_users
-    self.mentioned_users.each do |user|
-      CommentMailer.notify_callout(self,user) if user.username != self.author.username
-    end
-    already = self.mentioned_users.collect(&:uid)
-    uids = []
-    # notify note author, other commenters, and likers, but not those already @called out
-    (self.parent.comments.collect(&:uid) + [self.parent.uid] +
-      self.parent.likers.collect(&:uid)).uniq.each do |u|
 
+    notify_callout_users
+
+    already = self.mentioned_users.collect(&:uid) + [self.parent.uid]
+    uids = []
+    # notify other commenters, and likers, but not those already @called out
+    (self.parent.comments.collect(&:uid) + self.parent.likers.collect(&:uid)).uniq.each do |u|
       uids << u unless already.include?(u)
     end
-    DrupalUsers.find(:all, :conditions => ['uid IN (?)',uids]).each do |user|
-      if user.uid != current_user.uid &&
-        user.uid != self.uid &&
-        self.parent.uid != user.uid
 
-        CommentMailer.notify(user.user,self).deliver
-      end
-    end
+    notify_users(uids, current_user)
   end
 
   def answer_comment_notify(current_user)
@@ -123,23 +117,17 @@ class DrupalComment < ActiveRecord::Base
       CommentMailer.notify_answer_author(self.answer.author,self).deliver
     end
 
-    # notify mentioned users
-    self.mentioned_users.each do |user|
-      CommentMailer.notify_callout(self,user) if user.username != self.author.username
-    end
+    notify_callout_users
 
     already = self.mentioned_users.collect(&:uid) + [self.answer.uid]
     uids = []
-    # notify users who liked the answer and others involved in the conversation
+    # notify other answer commenter and users who liked the answer
+    # except mentioned users and answer author
     (self.answer.comments.collect(&:uid) + self.answer.likers.collect(&:uid)).uniq.each do |u|
       uids << u unless already.include?(u)
     end
 
-    DrupalUsers.find(:all, :conditions => ['uid IN (?)',uids]).each do |user|
-      if user.uid != current_user.uid
-        CommentMailer.notify(user.user,self).deliver
-      end
-    end
+    notify_users(uids, current_user)
   end
 
 end
